Restructure the whole project
This restructures the project in to three seperate CMake projects. This
is needed since every board should be build against a different MCU and
the hacks we were preforming to make that work were... not good.

In the new architecture, ever project builds its own test and
deployment binaries. The build.sh script will still orchetsrate builds
of all the binaries, but now it's performed via a bunch of different
CMake invocations, two for each of the three target boards.