Allow usage of uart_t directly
Prior to this commit, uart_t was available only as an opaque type when
including "uart.h" from the common library. However, this turns out to
be really inconvenient when we want to embed UART structures in other
locations as the client code would be responsible for juggling test
versus native implementations of the uart_t structure.